@@ -1,6 +1,6 @@
 import { Component, OnInit, ViewChild, ElementRef, OnDestroy } from '@angular/core';
 import { Observable, Subscription } from 'rxjs';
-import { finalize, take } from 'rxjs/operators';
+import { finalize } from 'rxjs/operators';
 import { FormBuilder, FormGroup } from '@angular/forms';
 import { UserFrameComponent } from '../user-frame/user-frame.component';
 
@@ -46,7 +46,7 @@ export class QuadrantComponent implements OnInit, OnDestroy {
   quadrantForm: FormGroup;
 
   constructor(private taskService: TaskServiceBase, 
-              private taskModifierServiceBase: TaskModifierServiceBase, 
+              private taskModifierService: TaskModifierServiceBase, 
               private formBuilder: FormBuilder, 
               private dragulaService: DragulaService, 
               private frameComponent: UserFrameComponent) {
@@ -69,6 +69,9 @@ export class QuadrantComponent implements OnInit, OnDestroy {
 
     // wire up data event
     var sub = this.frameComponent.dataReadyToLoad.subscribe(item => this.onDataReadyToLoad());
+    this.subscriptions.push(sub); // capture for destruction
+
+    var sub = this.taskModifierService.taskQuadrantUpdated.subscribe(item => this.onTaskQuadrantUpdated());
     this.subscriptions.push(sub); // capture for destruction
   }
 
@@ -140,7 +143,7 @@ export class QuadrantComponent implements OnInit, OnDestroy {
     var quadrantOld = source.id.substring(target.id.length - 1)
     var quadrantNew = target.id.substring(target.id.length - 1)
 
-    console.log("Element " + element.id + " moved (" + quadrantOld + "->" + quadrantNew + ")");
+    console.log("Requested move of element " + element.id + " (" + quadrantOld + "->" + quadrantNew + ")");
 
     // Get Dragula "drake" API reference (ng2-dragula doc refers to dragula)
     // https://github.com/bevacqua/dragula#readme
@@ -154,32 +157,17 @@ export class QuadrantComponent implements OnInit, OnDestroy {
   }
 
   private updateTask(taskId: string, targetQuadrant: string, dragulaDrake: any){
-    // get fresh task to work upon
-    this.taskService.getTask(taskId, this.selectedTaskList)
-    .pipe(take(1))
-    .subscribe((task: ITask) => 
-      {
-        // Update notes of fresh task
-        this.taskModifierServiceBase.setQuadrant(task, targetQuadrant);
+    this.taskModifierService.updateTaskQuadrant(this.taskService, taskId, this.selectedTaskList, targetQuadrant);
+  }
 
-        // Commit updated task notes via Google API
-        this.taskService.updateTask( task, this.selectedTaskList
-        ).then((task) => {
-          console.log("Task " + task.id + " successfully updated via API.");
-
-          // TODO: Optimize reload to remove flicker
-          // Update model with committed data
-          this.loadTasks(this.selectedTaskList);
-
-        }).catch((errorHandler) => {
-          console.log('Error in QuadrantComponent.onDrop: UpdateTask: ' + ((errorHandler == null || errorHandler.result == null) ? "undefined errorHandler" : errorHandler.result.error.message));
-        });
-      }
-    );
+  onTaskQuadrantUpdated() {
+    // TODO: Optimize reload to remove flicker
+    // Update model with committed data
+    this.loadTasks(this.selectedTaskList);
   }
 
   // Called by repeater to determine appropriate quadrant for each task
   quadrantMatch(task: ITask, quadrant:string): boolean {
-    return this.taskModifierServiceBase.checkQuadrantMatch(task, quadrant);
+    return this.taskModifierService.checkQuadrantMatch(task, quadrant);
   }
 }
