@@ -1,13 +1,51 @@
-import { Injectable } from '@angular/core';
+import { Injectable,Output, EventEmitter } from '@angular/core';
+import { take } from 'rxjs/operators';
+import { Subscription } from '../../../node_modules/rxjs';
+import { AutoUnsubscribe } from "ngx-auto-unsubscribe";
+
 import { TaskModifierServiceBase } from './task-modifier-service-base';
 import { ITask } from '../models/itask';
+import { TaskServiceBase } from './task-service-base';
 
+@AutoUnsubscribe({includeArrays: true})
 @Injectable({
   providedIn: 'root'
 })
 export class TaskModifierService implements TaskModifierServiceBase {
+  @Output() taskQuadrantUpdated: EventEmitter<any> = new EventEmitter();
+
+  // these subscriptions will be cleaned up by @AutoUnsubscribe
+  private subscriptions: Subscription[] = [];
 
   constructor() { }
+
+  updateTaskQuadrant(taskService: TaskServiceBase, taskId: string, taskListId: string, quadrantChar: string) {
+    var sub: Subscription;
+
+    // get fresh task to work upon
+    sub = taskService.getTask(taskId, taskListId)
+    .pipe(take(1))
+    .subscribe((task: ITask) => 
+      {
+        // Update notes of fresh task
+        this.setQuadrant(task, quadrantChar);
+
+        // Commit updated task notes via Google API
+        taskService.updateTask( task, taskListId
+        ).then((task) => {
+          console.log("Task " + task.id + " successfully updated via API.");
+          this.onTaskQuadrantUpdated();
+        }).catch((errorHandler) => {
+          console.log('Error in QuadrantComponent.onDrop: UpdateTask: ' + ((errorHandler == null || errorHandler.result == null) ? "undefined errorHandler" : errorHandler.result.error.message));
+        });
+      }
+    );
+    this.subscriptions.push(sub); // capture for destruction
+  }
+
+  private onTaskQuadrantUpdated() {
+    this.taskQuadrantUpdated.emit();
+  }
 
   setQuadrant(task: ITask, quadrantChar: string) {
     var tagPositionStart = -1;
